+++ Editor/AddressableImportSettings.cs
@@ -3,21 +3,23 @@ using UnityEditor;
 using UnityEditor.AddressableAssets;
 using System.Collections.Generic;
 using System.Linq;
+using System.IO;
 using UnityAddressableImporter.Helper;
 
 #if ODIN_INSPECTOR
 using Sirenix.OdinInspector;
 #endif
 
-[CreateAssetMenu(fileName = "AddressableImportSettings", menuName = "Addressable Assets/Import Settings", order = 50)]
 public class AddressableImportSettings : ScriptableObject
 {
-    public const string kDefaultConfigObjectName = "addressableimportsettings";
-    public const string kDefaultPath = "Assets/AddressableAssetsData/AddressableImportSettings.asset";
+    [Tooltip("Toggle rules enabled state")]
+    [SerializeField]
+    public bool rulesEnabled = true;
 
     [Tooltip("Creates a group if the specified group doesn't exist.")]
     public bool allowGroupCreation = false;
 
+    [Space]
     [Tooltip("Rules for managing imported assets.")]
 #if ODIN_INSPECTOR
     [ListDrawerSettings(HideAddButton = false,Expanded = false,DraggableItems = true,HideRemoveButton = false)]
@@ -58,19 +60,35 @@ public class AddressableImportSettings : ScriptableObject
         }
     }
 
-    public static AddressableImportSettings Instance
+    /// <summary>
+    /// Create AddressableImportSettings and add it to AddressableImportSettingsList
+    /// </summary>
+    [MenuItem("Assets/Create/Addressables/Import Settings", false, 50)]
+    public static void CreateAsset()
     {
-        get
+        string directoryPath = "Assets/";
+        string fileName = "AddressableImportSettings.asset";
+
+        foreach(var obj in Selection.GetFiltered(typeof(UnityEngine.Object), SelectionMode.Assets))
         {
-            AddressableImportSettings so;
-            // Try to locate settings via EditorBuildSettings.
-            if (EditorBuildSettings.TryGetConfigObject(kDefaultConfigObjectName, out so))
-                return so;
-            // Try to locate settings via path.
-            so = AssetDatabase.LoadAssetAtPath<AddressableImportSettings>(kDefaultPath);
-            if (so != null)
-                EditorBuildSettings.AddConfigObject(kDefaultConfigObjectName, so, true);
-            return so;
+            var assetPath = AssetDatabase.GetAssetPath(obj);
+            var assetDirectoryPath = AssetDatabase.IsValidFolder(assetPath) ? assetPath : Path.GetDirectoryName(assetPath);
+            if (AssetDatabase.IsValidFolder(assetDirectoryPath))
+            {
+                directoryPath = assetDirectoryPath;
+            }
         }
+        AddressableImportSettings settings = ScriptableObject.CreateInstance<AddressableImportSettings>();
+        var filePath = AssetDatabase.GenerateUniqueAssetPath(Path.Combine(directoryPath, fileName));
+        AssetDatabase.CreateAsset(settings, filePath);
+        Debug.LogFormat("Created AddressableImportSettings at path: {0}", filePath);
+
+        if (!AddressableImportSettingsList.Instance.SettingList.Contains(settings))
+        {
+            AddressableImportSettingsList.Instance.SettingList.Add(settings);
+        }
+
+        AssetDatabase.SaveAssets();
+        Selection.activeObject = settings;
     }
 }